URLパラメータをファイルに切り出し
